Boljsa arhitektura v relaciji okvir in Odsek

+++ api/gui/src/main/kotlin/gui/domain/Okvir.kt
@@ -1,11 +1,11 @@
 package gui.domain
 
 data class Okvir(
-    var x0: Double,
-    var x1: Double,
-    var y0: Double,
-    var y1: Double,
+    var x0: Int,
+    var x1: Int,
+    var y0: Int,
+    var y1: Int,
 ) {
-    val sirina: Double get() = x1 - x0
-    val visina: Double get() = y1 - y0
+    val sirina: Int get() = x1 - x0
+    val visina: Int get() = y1 - y0
 }
